stats: Rename stats.c to stats_prefix.c
